chore: improve images and chains
