+++ src/common/consts/path.ts
@@ -1,0 +1,1 @@
+export const RAW_FILES_PATH = 'https://raw.githubusercontent.com/cowprotocol/cow-sdk/refs/heads/main'
